Completed Thermocouple Input Software
Finished the Thermocouple Input program, it is written completely new in
thermo2.asm. Once we're okay with the changes, I'll combine the program
back into the original Thermocouple_Input.asm file.

Use the file Test_Thermo2.asm with the hardware to see if we are
accurately relaying the temperature.

I've hidden a lot of the helper subroutines in the program, so when we
use it in the main file, we only need two subroutines: "Initialize" and
"Read".

"Initialize" will set up all the ports so that we can accept data
"Read" will read from the ADC and automatically store the data to a
2-byte Temperature variable in Celsius.

Refer to the header file for more information