This is the main folder for the project report
The word document is the main report, I will be editing and revising the
formatting of it. If you are done your parts please label which part of
the report you have done in the title of a NEW DOCUMENT so I can copy it
over with the right format.